backend/requirements.txt
# Core Django
django>=5.0
djangorestframework
django-cors-headers

# Database & Models
Pillow>=10.0.0

# Authentication & Security
argon2-cffi>=23.1.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
PyJWT>=2.8.0

# API Documentation
drf-spectacular>=0.27.0

# Email & Notifications
django-templated-mail>=1.1.1

# Text Analysis (for AI service)
textblob>=0.17.1

# Utilities
python-decouple>=3.8
django-filter>=23.5
django-ratelimit>=4.1.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Development
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0